Branch wip-daffodil created.
update build.sbt and sbt version

See https://github.com/DFDLSchemas/Cobol/pull/1
There are comments on improving this schema there. The PR
was closed accidently, but is still accessible.

This schema is a first cut at some Cobol from daffodil.
It's not complete yet due to limitations in Daffodil.